Update template.less
The search button was not centered within the search box. This change adds an additional 5px of padding to the left of the search placeholder text, as well as adds 5px to the top and bottom padding of the search icon. This centers it within the box, and between the left edge of the box and the search placeholder text.